remove landing page & guides, it’s now API only
