+++ src/components/layout.js
@@ -1,15 +1,9 @@
 import React, { Component } from 'react'
-import { StaticQuery, Link } from 'gatsby'
+import { graphql, StaticQuery } from 'gatsby'
 
 import layoutStyles from './layout.module.css'
 import './layout.css'
 import Search from "./search"
-
-const ListLink = props => (
-  <li style={{ display: `inline-block`, marginRight: `1rem` }}>
-    <Link to={props.to}>{props.children}</Link>
-  </li>
-)
 
 class Layout extends Component {
   constructor(props) {
@@ -29,15 +23,7 @@ class Layout extends Component {
   render() {
     return <div>
       <header className={layoutStyles.header}>
-        <Link to='/' style={{ textShadow: `none`, backgroundImage: `none` }}>
-          <h3 style={{ display: `inline` }}>octokit/rest.js</h3>
-        </Link>
-
-        <ul style={{ listStyle: `none`, float: `right` }}>
-          <ListLink to='/'>Home</ListLink>
-          <ListLink to='/guides/'>Guides</ListLink>
-          <ListLink to='/api/'>API</ListLink>
-        </ul>
+        <h3 style={{ display: `inline` }}>octokit/rest.js</h3>
 
         <StaticQuery
           query={graphql`
