+++ src/pages/index.js
@@ -1,14 +1,175 @@
-import React from 'react'
-import Layout from '../components/layout'
-import { Link } from 'gatsby'
+import Prism  from 'prismjs'
 
-export default () => (
+import marked from 'marked'
+import React, { Component } from 'react'
+
+import Layout from '../components/layout'
+import { graphql } from 'gatsby'
+import apiStyles from '../components/api.module.css'
+
+class ApiSubMenu extends Component {
+  constructor(props) {
+    super(props)
+    this.showMenu = this.showMenu.bind(this)
+    this.isActive = this.isActive.bind(this)
+  }
+
+  showMenu() {
+    this.props.onUserInteraction(this.props.node.id)
+  }
+
+  isActive() {
+    return this.props.isActive(this.props.node.id)
+  }
+
+  render() {
+    return (
+      <li key={this.props.node.id}>
+        <a href={`#${this.props.node.id}`} onClick={this.showMenu} onFocus={this.showMenu} className={this.isActive() ? apiStyles.activelink : ""}>{this.props.node.name}</a>
+        <ol className={this.isActive() ? "" : apiStyles.subhidden}>
+          {this.props.node.methods.map((method) => {
+            return <li key={method.id}>
+              <a href={`#${method.id}`}>{method.name}</a>
+            </li>
+          })}
+        </ol>
+      </li>
+    )
+  }
+}
+
+class ApiMenu extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      menuActive: false,
+      activeSubMenu: null
+    }
+    this.toggleMenu       = this.toggleMenu.bind(this)
+    this.setActiveSubMenu = this.setActiveSubMenu.bind(this)
+    this.isActiveSubMenu  = this.isActiveSubMenu.bind(this)
+  }
+
+  toggleMenu() {
+    this.setState({
+      menuActive: !this.state.menuActive
+    })
+  }
+
+  setActiveSubMenu(subMenu) {
+    this.setState({
+      activeSubMenu: subMenu
+    })
+  }
+
+  isActiveSubMenu(subMenu) {
+    return this.state.activeSubMenu === subMenu
+  }
+
+  render() {
+    return (
+      <nav className={apiStyles.nav}>
+        <button type="button" onClick={this.toggleMenu}>Menu</button>
+        <h1 className={this.state.menuActive ? "" : apiStyles.hidden}>API</h1>
+        <ol className={this.state.menuActive ? "" : apiStyles.hidden}>
+          {this.props.data.staticMethods.edges.map(({ node }) => {
+            return <li key={node.id}>
+              <a href={`#octokit-${node.fields.idName}`}>{node.frontmatter.title}</a>
+            </li>
+          })}
+          {this.props.data.endpointScopes.edges.map(({ node }) => {
+            return <ApiSubMenu key={node.id} node={node} onUserInteraction={this.setActiveSubMenu} isActive={this.isActiveSubMenu}></ApiSubMenu>
+          })}
+        </ol>
+      </nav>
+    )
+  }
+}
+
+export default ({ data }) => (
   <Layout>
-    <div>
-      <h1>Home</h1>
-      <p>
-        Check out our <Link to='guides'>guides</Link> or <Link to='api'>API</Link>.
-      </p>
-    </div>
+    <ApiMenu data={data}></ApiMenu>
+    <main className={apiStyles.container}>
+      {data.staticMethods.edges.map(({ node }) => {
+        return <React.Fragment key={node.id}>
+          <h2>{node.frontmatter.title}</h2>
+          <div dangerouslySetInnerHTML={{ __html: node.html }} />
+        </React.Fragment>
+      })}
+
+{data.endpointScopes.edges.map(({ node }) => {
+        return <React.Fragment key={node.id}>
+          <h2 id={node.id}>{node.name}</h2>
+          {node.methods.map(method => {
+            return <React.Fragment key={method.id}>
+              <h3 id={method.id}>{method.name}</h3>
+              <div dangerouslySetInnerHTML={{ __html: marked(method.description) }} />
+              <h4>Parameters</h4>
+              <div className={apiStyles.table}>
+                <table>
+                  <thead>
+                    <tr>
+                      <th>name</th>
+                      <th>required</th>
+                      <th>description</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    {method.params.map(param => {
+                      return <tr key={param.name}>
+                        <td>{param.name}</td>
+                        <td>{param.required ? 'yes' : 'no'}</td>
+                        <td>{param.description}</td>
+                      </tr>
+                    })}
+                  </tbody>
+                </table>
+              </div>
+              <div className="gatsby-highlight" data-language="js">
+                <pre className="language-js"><code className="language-js" dangerouslySetInnerHTML={{ __html: Prism.highlight(method.example, Prism.languages.javascript, 'javascript') }} ></code></pre>
+              </div>
+            </React.Fragment>
+          })}
+        </React.Fragment>
+      })}
+    </main>
   </Layout>
 )
+
+export const query = graphql`
+  query {
+    staticMethods: allMarkdownRemark {
+      edges {
+        node {
+          id
+          frontmatter {
+            title
+          }
+          html
+          fields {
+            idName
+          }
+        }
+      }
+    }
+    endpointScopes: allOctokitApiGroup {
+      edges {
+        node {
+          id
+          name
+          methods {
+            id
+            name
+            description
+            example
+            params {
+              name
+              required
+              description
+            }
+          }
+        }
+      }
+    }
+  }
+`
